extracted level up screen
